README.md
This project is a part of my FreeCodeCamp certification process. It's a Random Quote Machine built using React, a frontend framework, and an external API to fetch and display random quotes. The project also features a share button for Twitter and a visually appealing design that enhances the user experience.